This revision adds the solution to the classical Invoice problem using Go4 Design-Patterns and others using Java 6. It includes test cases and test data as well.
http://www.coderanch.com/t/495510/java/java/Case-Study-Sales-Tax

Basic sales tax is applicable at a rate of 10% on all goods, except books, 
food, and medical products that are exempt. Import duty is an additional 
sales tax applicable on all imported goods at a rate of 5%, with no 
exemptions. 

When I purchase items I receive a receipt which lists the name of all the 
items and their price (including tax), finishing with the total cost of the 
items, and the total amounts of sales taxes paid. The rounding rules for 
sales tax are that for a tax rate of n%, a shelf price of p contains 
(np/100 rounded up to the nearest 0.05) amount of sales tax. 

Write an application that prints out the receipt details for these shopping 
baskets...